RavenDB-18164 - add support for filter in sharded queries
